rgw: make appending Content-Length in 204 and 304 responses controllable.
This commit introduces a new configurable rgw_always_print_content_length
to let operator decide whether RadosGW complies to RFC 7230 (a part of
the HTTP specification) or violates it but follows the Swift behavior.

Having this configurable is essential to pass the RefStack. That is, at
the moment  Tempest enforces all implementations of the OpenStack Object
Storage API to violate RFC 7230 [1]. Two independent fixes for this issue
have been refused [2][3].

[1] https://bugs.launchpad.net/tempest/+bug/1536251
[2] https://review.openstack.org/#/c/272062
[3] https://review.openstack.org/#/c/275652

Fixes: http://tracker.ceph.com/issues/16602
Signed-off-by: Radoslaw Zarzynski <rzarzynski@mirantis.com>