ceph.spec.in: drop trouble-making divider comments
We have some %post scriptlets that fire up a shell just to execute a single
command like this:

%postun -n librados2
/sbin/ldconfig

Normally, it would be possible to optimize these as:

%postun -n librados2 -p /sbin/ldconfig

Except that, due to an RPM bug, there must not be *any* content (not even comments)
between the %postun and the next section.

See https://bugzilla.redhat.com/show_bug.cgi?id=1003962#c0 for details.

Signed-off-by: Nathan Cutler <ncutler@suse.com>